+++ core/models.py
@@ -14,29 +14,6 @@ class Base(models.Model):
         abstract = True
 
 
-class Produto(Base):
-    SKU = models.CharField('SKU', max_length=10)
-    NomeProduto = models.CharField('Nome do Produto', max_length=100)
-    categoria = models.ForeignKey('core.CategoriaProduto', verbose_name='Categoria', on_delete=models.CASCADE)
-    estoque = models.IntegerField('Estoque', default=0)
-
-    class Meta:
-        verbose_name = 'Produto'
-        verbose_name_plural = 'Produtos'
-
-    def __str__(self):
-        return self.SKU
-
-
-class CategoriaProduto(Base):
-    categoria = models.CharField('Categoria', max_length=20)
-
-    class Meta:
-        verbose_name = 'Categoria'
-        verbose_name_plural = 'Categorias'
-
-    def __str__(self):
-        return self.categoria
 
 
 class Conta(Base):
