+++ unicatech/urls.py
@@ -24,4 +24,5 @@ urlpatterns = [
     path('', include('core.urls')),
     path('', include('Vendas.urls')),
     path('', include('Compras.urls')),
+    path('', include('Produtos.urls')),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
